refactor: move slope calculation into form hook
