+++ src/lib/contracts/locking/useRelockMento.ts
@@ -9,14 +9,11 @@ import { LockingABI } from "@/lib/abi/Locking";
 import { Address } from "viem";
 import { WriteContractErrorType } from "wagmi/actions";
 import useTokens from "../useTokens";
-import useLockingWeek from "./useLockingWeek";
-import { MAX_LOCKING_DURATION_WEEKS } from "@/components/_shared/mento-lock/constants";
-import { differenceInWeeks } from "date-fns";
 import { ExtendedLock } from "@/lib/hooks/useLockInfo";
 interface RelockMentoParams {
   newDelegate?: Address;
   newAmount?: bigint;
-  newExpirationDate: Date;
+  newSlope: number;
   newCliff?: number;
   onSuccess?: () => void;
   onError?: (error?: WriteContractErrorType) => void;
@@ -28,37 +25,19 @@ const useRelockMento = ({
   newAmount,
   newCliff,
   newDelegate,
-  newExpirationDate,
   onError,
   onSuccess,
+  newSlope,
 }: RelockMentoParams) => {
   const contracts = useContracts();
   const { veMentoBalance } = useTokens();
-  const { currentWeek: currentLockingWeek } = useLockingWeek();
+
   const {
     writeContract,
     isPending: isAwaitingUserSignature,
     data,
     ...restWrite
   } = useWriteContract();
-
-  const newSlope = React.useMemo(() => {
-    if (!newExpirationDate) return 0;
-
-    const currentSlope = lock?.slope;
-    const weeksPassed = Number(currentLockingWeek) - lock?.time;
-    const weeksAdded = differenceInWeeks(newExpirationDate, lock?.expiration);
-    return Math.min(
-      currentSlope - weeksPassed + weeksAdded,
-      MAX_LOCKING_DURATION_WEEKS,
-    );
-  }, [
-    currentLockingWeek,
-    lock?.expiration,
-    lock?.slope,
-    lock?.time,
-    newExpirationDate,
-  ]);
 
   const lockingArgs = React.useMemo(() => {
     return [
